34d24e93b540b11fa11684226978c4405ea23896 Fix overflow on 16 bit systems (Pieter Wuille)

Pull request description:

  Fixes #15

Top commit has no ACKs.

Tree-SHA512: 8c7c07b6f14835972811b1507034bba36f99a55fd7e5cfe0bd43049610925c3e9ae01d8b9371d9a17245e7e2796a652e822564d66c146323e402505131524e69

+++ ctaes.c
@@ -25,7 +25,7 @@
 static void LoadByte(AES_state* s, unsigned char byte, int r, int c) {
     int i;
     for (i = 0; i < 8; i++) {
-        s->slice[i] |= (byte & 1) << (r * 4 + c);
+        s->slice[i] |= (uint16_t)(byte & 1) << (r * 4 + c);
         byte >>= 1;
     }
 }
@@ -255,7 +255,7 @@ static void SubBytes(AES_state *s, int inv) {
     }
 }
 
-#define BIT_RANGE(from,to) (((1 << ((to) - (from))) - 1) << (from))
+#define BIT_RANGE(from,to) ((uint16_t)((1 << ((to) - (from))) - 1) << (from))
 
 #define BIT_RANGE_LEFT(x,from,to,shift) (((x) & BIT_RANGE((from), (to))) << (shift))
 #define BIT_RANGE_RIGHT(x,from,to,shift) (((x) & BIT_RANGE((from), (to))) >> (shift))
